refactor(graphql): rename pageInfoResolver tests' file, use *.test.ts
Renamed resolvers.test.js → pageInfoResolver.test.ts as the tests were
really only about pageInfoResolver and nothing else, and removed one
nesting level as unnecessary from the tests.

refs US-171
